Rename pinecone_hook to pinecone as per AIP-21
