+++ airflow/providers/pinecone/hooks/pinecone.py
